+++ app/src/main/java/com/bs/sriwilis/nasabah/data/model/CartOrder.kt
@@ -7,7 +7,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class CartOrder (
     val kategori: String,
-    val berat_perkiraan: Int,
+    val berat_perkiraan: Float,
     val harga_perkiraan: Double,
     val gambar: String?,
 ) : Parcelable
